Renaming branch to specify Community portion
